Commented and debugged

+++ src/model/Model.java
@@ -86,7 +86,7 @@ public class Model implements ColorIndexObservable{
 
 	public void addTurtle(double ID) {
 		Turtle t = new Turtle (XHome, YHome, Color.BLUE, ID, TURTLE_0);
-		allTurtles.put((double) t.getValue(), t);
+		allTurtles.put( t.getValue(), t);
 		activeTurtles.add(t);
 	}
